backend/requirements.txt
Django
djangorestframework
pygments
mpmath

#
httpie
pre-commit
